Added navigation to and from about screen
