+++ app/src/main/java/com/teamb/moviespoc/navigation/NavGraph.kt
@@ -5,6 +5,7 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import com.teamb.moviespoc.AnimatedSplashScreen
+import com.teamb.moviespoc.ui.about.About
 import com.teamb.moviespoc.ui.home.Home
 import com.teamb.moviespoc.ui.welcome.Authenticate
 
@@ -25,10 +26,13 @@ fun SetUpNavGraph(navController: NavHostController){
             AnimatedSplashScreen(navigationActions.navigateToWelcome)
         }
         composable(route = Screen.Home.route) {
-            Home()
+            Home(navigationActions.navigateToAbout)
         }
         composable(route = Screen.Welcome.route) {
             Authenticate(navigationActions.navigateToHome, navigationActions.navigateToHome)
+        }
+        composable(route = Screen.About.route) {
+            About(navigationActions.navigateToHome)
         }
     }
 }
